# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mentis Semear API",
        "description": "API do monólito Mentis Semear (autenticação JWT, workshops, check-ins, avaliações e diário).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Mesma origem (dev e produção)"
        }
    ],
    "paths": {
        "/api/assessments": {
            "get": {
                "tags": [
                    "Assessments"
                ],
                "summary": "Lista avaliações (paginado)",
                "operationId": "d88c31516fb8559c3559254bcb23a6ef",
                "parameters": [
                    {
                        "name": "workshop_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Assessment"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Assessments"
                ],
                "summary": "Registra uma avaliação anônima",
                "operationId": "0cc4ab6ca8a3e9dc31d6cf7476849c32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workshop_id",
                                    "score"
                                ],
                                "properties": {
                                    "workshop_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 9,
                                        "maximum": 10,
                                        "minimum": 0
                                    },
                                    "suggestions": {
                                        "type": "string",
                                        "example": "Muito bom!",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Assessment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/public/assessments": {
            "post": {
                "tags": [
                    "Público"
                ],
                "summary": "Avaliação pública anônima (sem autenticação)",
                "operationId": "7f083c5cd70b2571f59d3521856eef9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "workshop_id",
                                    "score"
                                ],
                                "properties": {
                                    "workshop_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "type": "integer",
                                        "example": 9,
                                        "maximum": 10,
                                        "minimum": 0
                                    },
                                    "suggestions": {
                                        "type": "string",
                                        "example": "Muito bom!",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Avaliação registrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Assessment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                }
            }
        },
        "/api/assessments/{assessment}": {
            "get": {
                "tags": [
                    "Assessments"
                ],
                "summary": "Detalha uma avaliação",
                "operationId": "6a8c7c077e00528d519e3fb0cf93c3d8",
                "parameters": [
                    {
                        "name": "assessment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avaliação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Assessment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Assessments"
                ],
                "summary": "Atualiza uma avaliação (campos parciais)",
                "operationId": "be0427370843599db02262ce227117ff",
                "parameters": [
                    {
                        "name": "assessment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "score": {
                                        "type": "integer",
                                        "maximum": 10,
                                        "minimum": 0
                                    },
                                    "suggestions": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Assessment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrada"
                    },
                    "422": {
                        "description": "Falha de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assessments"
                ],
                "summary": "Remove uma avaliação",
                "operationId": "ee6b35faf717390180c65662141cc1bf",
                "parameters": [
                    {
                        "name": "assessment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Acesso restrito ao perfil"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Assessment": {
                "properties": {
                    "id": {
                        "description": "OpenAPI schema only — never instantiated.",
                        "type": "integer",
                        "example": 1
                    },
                    "workshop_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "score": {
                        "type": "integer",
                        "example": 9,
                        "maximum": 10,
                        "minimum": 0
                    },
                    "suggestions": {
                        "type": "string",
                        "example": "Muito bom!",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Envie o token no header: Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}